David (Js & Py) Solution
This was a fun exercise, hopefully I did everything in a clean and well optimized manner. There are defiantly better ways of going at this.

I decided to take on the task to do both languages to see what I can come up with for solutions with both of them. I would love to go over my train of thought during the interview and go over better solutions then what I came to.

Thank you for this challenge :smile: